split verify signatures test
