+++ tests/test_vectors/CMakeLists.txt
@@ -7,6 +7,7 @@
 addtest(test_vectors_test
     ssz_test.cpp
     test_vectors_test.cpp
+    verify_signatures_test.cpp
 )
 target_link_libraries(test_vectors_test
     blockchain
